requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
msal>=1.20.0
requests>=2.25.0
openai>=1.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
PyYAML>=6.0
sqlite-utils>=3.0.0
